Add useTableSort hook to handle sorting for machine and KVM lists.
